pyproject.toml
[tool.poetry]
name = "deezer-to-spotify"
version = "0.1.0"
description = ""
authors = ["pppp56"]
readme = "README.md"
packages = [{include = "deezer_to_spotify"}]

[tool.poetry.dependencies]
python = "^3.9"
deezer-oauth-cli = "^1.1.2"
deezer-python = "^6.1.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.linting.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
